refactor(debate): absorb dialogue-prompts.ts into DebatePromptBuilder — Phase 4
Move all review-specific prompt builders from src/review/dialogue-prompts.ts
into DebatePromptBuilder, unifying plan and review debate prompts behind a
single class. This eliminates the parallel prompt universe that required
every fix (persona injection, format separation, labeling) to be
replicated in two places.

Changes:
- Add 4 public methods to DebatePromptBuilder: buildReviewPrompt,
  buildReReviewPrompt, buildResolverPrompt, buildReResolverPrompt
- Add 4 private helpers: buildResolverFraming, buildVoteTallyLine,
  buildLabeledProposalsSection, buildLabeledCritiquesSection
- Move DebateResolverContext type from dialogue-prompts.ts to debate/types.ts
- Migrate dialogue.ts to use shared promptBuilder instance
- Export ReviewStoryContext and DebateResolverContext from debate barrel
- Delete dialogue-prompts.ts and its test file (23 tests migrated)
