+++ src/debate/index.ts
@@ -7,7 +7,7 @@ export { _debateSessionDeps, resolveDebaterModel } from "./session-helpers";
 export type { DebateSessionOptions, ResolverContextInput, ResolveOutcome } from "./session-helpers";
 export { majorityResolver, synthesisResolver, judgeResolver } from "./resolvers";
 export { DebatePromptBuilder } from "./prompt-builder";
-export type { StageContext, PromptBuilderOptions } from "./prompt-builder";
+export type { StageContext, PromptBuilderOptions, ReviewStoryContext } from "./prompt-builder";
 export { PERSONA_FRAGMENTS, buildDebaterLabel, buildPersonaBlock, resolvePersonas } from "./personas";
 export type { DebaterPersona } from "./types";
 export type {
@@ -19,6 +19,7 @@ export type {
   Proposal,
   Rebuttal,
   ResolverConfig,
+  DebateResolverContext,
   ResolverType,
   SessionMode,
 } from "./types";
