+++ src/debate/prompt-builder.ts
@@ -6,10 +6,18 @@
  * method per prompt phase (proposal, critique, rebuttal, synthesis, judge, close).
  */
 
+import type { ReviewFinding } from "../plugins/types";
 import { PERSONA_FRAGMENTS } from "./personas";
-import type { Debater, Proposal, Rebuttal } from "./types";
+import type { DebateResolverContext, Debater, Proposal, Rebuttal } from "./types";
 
 // ─── Interfaces ───────────────────────────────────────────────────────────────
+
+/** Story fields required by review-specific prompt methods. */
+export interface ReviewStoryContext {
+  id: string;
+  title: string;
+  acceptanceCriteria: string[];
+}
 
 export interface StageContext {
   /** Task description — spec, codebase context, constraints. No output format. */
@@ -164,5 +172,147 @@ ${this.stageContext.outputFormat}`;
 
   private buildDebaterLabel(debater: Debater): string {
     return debater.persona ? `${debater.agent} (${debater.persona})` : debater.agent;
+  }
+
+  // ─── Review-specific public methods (Phase 4) ─────────────────────────────
+
+  /** Standard (non-debate) review prompt. */
+  buildReviewPrompt(diff: string, story: ReviewStoryContext): string {
+    const criteria = story.acceptanceCriteria.map((c) => `- ${c}`).join("\n");
+    return [
+      `Review the following code diff for story ${story.id}: ${story.title}`,
+      "",
+      "## Acceptance Criteria",
+      criteria,
+      "",
+      "## Diff",
+      diff,
+      "",
+      "Also flag any changes in the diff not required by the acceptance criteria above as out-of-scope findings.",
+      "Respond with JSON: { passed: boolean, findings: [...], findingReasoning: { [id]: string } }",
+    ].join("\n");
+  }
+
+  /** Re-review prompt — references previous findings and updated diff. */
+  buildReReviewPrompt(updatedDiff: string, previousFindings: ReviewFinding[]): string {
+    const findingsList =
+      previousFindings.length > 0 ? previousFindings.map((f) => `- ${f.ruleId}: ${f.message}`).join("\n") : "(none)";
+    return [
+      "This is a follow-up re-review. Please review the updated diff below.",
+      "",
+      "## Previous Findings",
+      findingsList,
+      "",
+      "## Updated Diff",
+      updatedDiff,
+      "",
+      "Respond with JSON: { passed: boolean, findings: [...], findingReasoning: { [id]: string }, deltaSummary: string }",
+      "deltaSummary should describe which previous findings are resolved vs still present.",
+    ].join("\n");
+  }
+
+  /** Debate resolver prompt — varies by resolver type. */
+  buildResolverPrompt(
+    proposals: Array<{ debater: string; output: string }>,
+    critiques: string[],
+    diff: string,
+    story: ReviewStoryContext,
+    resolverContext: DebateResolverContext,
+  ): string {
+    const criteria = story.acceptanceCriteria.map((c) => `- ${c}`).join("\n");
+    const framing = this.buildResolverFraming(resolverContext);
+    const voteTally = this.buildVoteTallyLine(resolverContext);
+    const proposalsSection = this.buildLabeledProposalsSection(proposals);
+    const critiquesSection = this.buildLabeledCritiquesSection(critiques);
+
+    return [
+      framing,
+      "",
+      `## Story ${story.id}: ${story.title}`,
+      "",
+      "## Acceptance Criteria",
+      criteria,
+      "",
+      "## Debater Proposals",
+      proposalsSection,
+      critiquesSection,
+      "",
+      "## Diff",
+      diff,
+      voteTally,
+      "",
+      "Respond with JSON: { passed: boolean, findings: [...], findingReasoning: { [id]: string } }",
+    ]
+      .filter((line) => line !== undefined)
+      .join("\n");
+  }
+
+  /** Debate re-review prompt — references previous findings + updated proposals. */
+  buildReResolverPrompt(
+    proposals: Array<{ debater: string; output: string }>,
+    critiques: string[],
+    updatedDiff: string,
+    previousFindings: ReviewFinding[],
+    resolverContext: DebateResolverContext,
+  ): string {
+    const framing = this.buildResolverFraming(resolverContext);
+    const findingsList =
+      previousFindings.length > 0 ? previousFindings.map((f) => `- ${f.ruleId}: ${f.message}`).join("\n") : "(none)";
+    const proposalsSection = this.buildLabeledProposalsSection(proposals);
+    const critiquesSection = this.buildLabeledCritiquesSection(critiques);
+
+    return [
+      `${framing} This is a re-review after implementer changes.`,
+      "",
+      "## Previous Findings",
+      findingsList,
+      "",
+      "## Updated Debater Proposals",
+      proposalsSection,
+      critiquesSection,
+      "",
+      "## Updated Diff",
+      updatedDiff,
+      "",
+      "Respond with JSON: { passed: boolean, findings: [...], findingReasoning: { [id]: string }, deltaSummary: string }",
+      "deltaSummary should describe which previous findings are resolved vs still present.",
+    ]
+      .filter((line) => line !== undefined)
+      .join("\n");
+  }
+
+  // ─── Review-specific private helpers ──────────────────────────────────────
+
+  private buildResolverFraming(ctx: DebateResolverContext): string {
+    switch (ctx.resolverType) {
+      case "majority-fail-closed":
+      case "majority-fail-open":
+        return "You are the authoritative reviewer resolving a debate. A preliminary vote was taken — see tally below. Verify disputed findings using tools (READ files, GREP for usage) and give your final verdict.";
+      case "synthesis":
+        return "You are a synthesis reviewer. Synthesize the debater proposals into a single, coherent, tool-verified verdict. Use READ and GREP to verify claims before ruling.";
+      case "custom":
+        return "You are the judge. Evaluate the debater proposals independently. Verify claims with tools (READ, GREP) and give your final authoritative verdict.";
+      default:
+        return "You are the reviewer. Evaluate the debater proposals and give your final authoritative verdict.";
+    }
+  }
+
+  private buildVoteTallyLine(ctx: DebateResolverContext): string {
+    if (!ctx.majorityVote) return "";
+    const { passCount, failCount } = ctx.majorityVote;
+    const failOpenNote =
+      ctx.resolverType === "majority-fail-open"
+        ? " (unparseable proposals count as pass)"
+        : " (unparseable proposals count as fail)";
+    return `\n\nThe preliminary majority vote is: **${passCount} passed, ${failCount} failed**${failOpenNote}. Verify the failing findings with tools before giving your authoritative verdict.`;
+  }
+
+  private buildLabeledProposalsSection(proposals: Array<{ debater: string; output: string }>): string {
+    return proposals.map((p) => `### ${p.debater}\n${p.output}`).join("\n\n");
+  }
+
+  private buildLabeledCritiquesSection(critiques: string[]): string {
+    if (critiques.length === 0) return "";
+    return `\n\n## Critiques\n${critiques.map((c, i) => `### Critique ${i + 1}\n${c}`).join("\n\n")}`;
   }
 }
